add test annotation singleton fix not found tikv-client-java-2.0-SNAPSHOT.jar
